@@ -83,72 +83,72 @@
     "packaging": "jar"
   },
   "dependencies": {
-    "@angular/common": "16.2.11",
-    "@angular/compiler": "16.2.11",
-    "@angular/core": "16.2.11",
-    "@angular/forms": "16.2.11",
-    "@angular/localize": "16.2.11",
-    "@angular/platform-browser": "16.2.11",
-    "@angular/platform-browser-dynamic": "16.2.11",
-    "@angular/router": "16.2.11",
-    "@fortawesome/angular-fontawesome": "0.13.0",
-    "@fortawesome/fontawesome-svg-core": "6.4.2",
-    "@fortawesome/free-solid-svg-icons": "6.4.2",
-    "@ng-bootstrap/ng-bootstrap": "15.1.2",
+    "@angular/common": "17.0.6",
+    "@angular/compiler": "17.0.6",
+    "@angular/core": "17.0.6",
+    "@angular/forms": "17.0.6",
+    "@angular/localize": "17.0.6",
+    "@angular/platform-browser": "17.0.6",
+    "@angular/platform-browser-dynamic": "17.0.6",
+    "@angular/router": "17.0.6",
+    "@fortawesome/angular-fontawesome": "0.14.0",
+    "@fortawesome/fontawesome-svg-core": "6.5.1",
+    "@fortawesome/free-solid-svg-icons": "6.5.1",
+    "@ng-bootstrap/ng-bootstrap": "16.0.0",
     "@ngx-translate/core": "15.0.0",
     "@ngx-translate/http-loader": "8.0.0",
     "@popperjs/core": "2.11.8",
     "bootstrap": "5.3.2",
     "dayjs": "1.11.10",
-    "ngx-infinite-scroll": "16.0.0",
+    "ngx-infinite-scroll": "17.0.0",
     "rxjs": "7.8.1",
     "tslib": "2.6.2",
-    "zone.js": "0.13.3"
+    "zone.js": "0.14.2"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "16.0.1",
-    "@angular-builders/jest": "16.0.1",
-    "@angular-devkit/build-angular": "16.2.9",
-    "@angular-eslint/eslint-plugin": "16.2.0",
-    "@angular/cli": "16.2.9",
-    "@angular/compiler-cli": "16.2.11",
-    "@angular/service-worker": "16.2.11",
-    "@types/jest": "29.5.7",
-    "@types/node": "18.18.8",
-    "@typescript-eslint/eslint-plugin": "6.9.1",
-    "@typescript-eslint/parser": "6.9.1",
+    "@angular-builders/custom-webpack": "17.0.0",
+    "@angular-builders/jest": "17.0.0",
+    "@angular-devkit/build-angular": "17.0.6",
+    "@angular-eslint/eslint-plugin": "17.1.1",
+    "@angular/cli": "17.0.6",
+    "@angular/compiler-cli": "17.0.6",
+    "@angular/service-worker": "17.0.6",
+    "@types/jest": "29.5.11",
+    "@types/node": "18.19.3",
+    "@typescript-eslint/eslint-plugin": "6.13.2",
+    "@typescript-eslint/parser": "6.13.2",
     "browser-sync": "2.29.3",
     "browser-sync-webpack-plugin": "2.3.0",
     "buffer": "6.0.3",
     "concurrently": "8.2.2",
     "copy-webpack-plugin": "11.0.0",
-    "cypress": "13.4.0",
+    "cypress": "13.6.1",
     "cypress-audit": "1.1.0",
-    "eslint": "8.52.0",
-    "eslint-config-prettier": "9.0.0",
+    "eslint": "8.55.0",
+    "eslint-config-prettier": "9.1.0",
     "eslint-plugin-cypress": "2.15.1",
     "eslint-webpack-plugin": "4.0.1",
     "folder-hash": "4.0.4",
-    "generator-jhipster": "8.0.0",
+    "generator-jhipster": "8.1.0",
     "husky": "8.0.3",
     "jest": "29.7.0",
     "jest-date-mock": "1.0.8",
     "jest-environment-jsdom": "29.7.0",
     "jest-junit": "16.0.0",
-    "jest-preset-angular": "13.1.2",
+    "jest-preset-angular": "13.1.4",
     "jest-sonar": "0.2.16",
-    "lighthouse": "11.2.0",
-    "lint-staged": "15.0.2",
+    "lighthouse": "11.3.0",
+    "lint-staged": "15.2.0",
     "merge-jsons-webpack-plugin": "2.0.1",
-    "prettier": "3.0.3",
-    "prettier-plugin-java": "2.3.1",
-    "prettier-plugin-packagejson": "2.4.6",
+    "prettier": "3.1.0",
+    "prettier-plugin-java": "2.5.0",
+    "prettier-plugin-packagejson": "2.4.7",
     "rimraf": "5.0.5",
-    "swagger-ui-dist": "5.9.1",
+    "swagger-ui-dist": "5.10.3",
     "ts-jest": "29.1.1",
-    "typescript": "5.1.6",
-    "wait-on": "7.0.1",
-    "webpack-bundle-analyzer": "4.9.1",
+    "typescript": "5.2.2",
+    "wait-on": "7.2.0",
+    "webpack-bundle-analyzer": "4.10.1",
     "webpack-merge": "5.10.0",
     "webpack-notifier": "1.15.0"
   },
