@@ -9,11 +9,13 @@ import { ProfileService } from './profile.service';
   standalone: true,
   selector: 'jhi-page-ribbon',
   template: `
-    <div class="ribbon" *ngIf="ribbonEnv$ | async as ribbonEnv">
-      <a href="" jhiTranslate="global.ribbon.{{ ribbonEnv }}">{{ { dev: 'Development' }[ribbonEnv] || '' }}</a>
-    </div>
+    @if (ribbonEnv$ | async; as ribbonEnv) {
+      <div class="ribbon">
+        <a href="" jhiTranslate="global.ribbon.{{ ribbonEnv }}">{{ { dev: 'Development' }[ribbonEnv] || '' }}</a>
+      </div>
+    }
   `,
-  styleUrls: ['./page-ribbon.component.scss'],
+  styleUrl: './page-ribbon.component.scss',
   imports: [SharedModule],
 })
 export default class PageRibbonComponent implements OnInit {
