@@ -7,7 +7,7 @@ import SharedModule from 'app/shared/shared.module';
   standalone: true,
   imports: [SharedModule],
   templateUrl: './password-strength-bar.component.html',
-  styleUrls: ['./password-strength-bar.component.scss'],
+  styleUrl: './password-strength-bar.component.scss',
 })
 export default class PasswordStrengthBarComponent {
   colors = ['#F00', '#F90', '#FF0', '#9F0', '#0F0'];
