@@ -55,7 +55,7 @@ export class DataUtils {
    * @param editForm the form group where the input field is located
    * @param field the field name to set the file's 'base 64 data' on
    * @param isImage boolean representing if the file represented by the event is an image
-   * @returns an observable that loads file to form field and completes if sussessful
+   * @returns an observable that loads file to form field and completes if successful
    *      or returns error as FileLoadError on failure
    */
   loadFileToForm(event: Event, editForm: FormGroup, field: string, isImage: boolean): Observable<void> {
