@@ -123,7 +123,7 @@ export class OperationComponent implements OnInit {
   }
 
   protected fillComponentAttributesFromResponseBody(data: IOperation[] | null): IOperation[] {
-    // If there is previus link, data is a infinite scroll pagination content.
+    // If there is previous link, data is a infinite scroll pagination content.
     if ('prev' in this.links) {
       const operationsNew = this.operations ?? [];
       if (data) {
