@@ -1,16 +1,16 @@
 import { enableProdMode } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { bootstrapApplication } from '@angular/platform-browser';
+import { appConfig } from './app/app.config';
+import AppComponent from './app/app.component';
 
 import { DEBUG_INFO_ENABLED } from './app/app.constants';
-import { AppModule } from './app/app.module';
 
 // disable debug data on prod profile to improve performance
 if (!DEBUG_INFO_ENABLED) {
   enableProdMode();
 }
 
-platformBrowserDynamic()
-  .bootstrapModule(AppModule, { preserveWhitespaces: true })
+bootstrapApplication(AppComponent, appConfig)
   // eslint-disable-next-line no-console
   .then(() => console.log('Application started'))
   .catch(err => console.error(err));
